Separate Nginx reload from certificate renewal
The following entry was found in /etc/cron.d/certbot on Debian 9.3:

    0 */12 * * * root test -x /usr/bin/certbot -a \! -d
    /run/systemd/system && perl -e 'sleep int(rand(3600))' && certbot -q
    renew

This entry performs a certifcate renewal attempt approximately every 12
hours. The exact time of renewal is chosen randomly. Since there is no
renewal hook in this entry to reload Nginx, Nginx continues to serve the
website with the old certificate which leads to connectivity problems
after the old certificate expires.

Therefore update our crontab entry so that it reloads Nginx once
everyday to ensure that it uses the most recently updated certificates.